loan endpoints implemented

+++ api/src/main/java/cl/torokoko/prestabanco/entity/LoanType.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 
 import java.util.Date;
+import java.util.List;
 
 @Entity
 @Table(name = "loan_type")
@@ -19,14 +20,14 @@ public class LoanType {
 
     @Getter @Setter private String name;
     @Column(name = "max_term")
-    @Getter @Setter private int maxTerm;
+    @Getter @Setter private float maxTerm;
 
     @Column(name = "min_rate")
-    @Getter @Setter private int minRate;
+    @Getter @Setter private float minRate;
 
     @Column(name = "max_rate")
-    @Getter @Setter private int maxRate;
+    @Getter @Setter private float maxRate;
 
     @Column(name = "max_funding")
-    @Getter @Setter private int maxFunding;
+    @Getter @Setter private float maxFunding;
 }
